misc/HtmlTestSuite/results/test-0170.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>Test parsing of the :host/host() pseudo-classes</title>
<link href="https://drafts.csswg.org/css-scoping/#selectordef-host-function" rel="help"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/css/support/parsing-testcommon.js"/>
<script>
  test_valid_selector(':host');
  test_valid_selector(':host(.a)');
  test_valid_selector(':host(div.a)');

  test_invalid_selector(':host()');
  test_invalid_selector(':host(.a, .b)');
  test_invalid_selector(':host(.a + .b)');
  test_invalid_selector(':host(.a + .b, #c &gt; #d)');
</script>
</head><body/></html>
